requirements.txt
alembic==1.4.1
Click==7.0
fake-rpigpio==0.1.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dateutil==2.8.1
python-editor==1.0.4
RPi.GPIO==0.7.0
RPi.version==1.2
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
